requirements.txt
pandas==1.3.3
numpy==1.21.1
Pillow==8.2.0
opencv-python==4.5.3
random2==1.0.1
mido==1.2.10
Flask==2.1.0
